+++ selfdrive/ui/qt/offroad/settings.h
@@ -1,8 +1,5 @@
 #pragma once
 
-#include <map>
-#include <string>
-
 #include <QButtonGroup>
 #include <QFrame>
 #include <QLabel>
@@ -56,26 +53,6 @@ private slots:
 private:
   Params params;
   ButtonControl *pair_device;
-};
-
-class TogglesPanel : public ListWidget {
-  Q_OBJECT
-public:
-  explicit TogglesPanel(SettingsWindow *parent);
-  void showEvent(QShowEvent *event) override;
-
-public slots:
-  void expandToggleDescription(const QString &param);
-
-private slots:
-  void updateState(const UIState &s);
-
-private:
-  Params params;
-  std::map<std::string, ParamControl*> toggles;
-  ButtonParamControl *long_personality_setting;
-
-  void updateToggles();
 };
 
 class SoftwarePanel : public ListWidget {
